Move convergence benchmark functionality to separate module
