+++ benchmarks/definition/config.py
@@ -1,10 +1,10 @@
-"""Benchmark configurations."""
+"""Basic benchmark configuration."""
 
 import time
 from abc import ABC
 from collections.abc import Callable
 from datetime import datetime, timedelta, timezone
-from typing import Any, Generic, TypeVar
+from typing import Generic, TypeVar
 
 from attrs import define, field
 from attrs.validators import instance_of
@@ -25,20 +25,6 @@ class BenchmarkSettings(ABC, BenchmarkSerialization):
 
 
 BenchmarkSettingsType = TypeVar("BenchmarkSettingsType", bound=BenchmarkSettings)
-
-
-@define(frozen=True)
-class ConvergenceBenchmarkSettings(BenchmarkSettings):
-    """Benchmark configuration for recommender convergence analyses."""
-
-    batch_size: int = field(validator=instance_of(int))
-    """The recommendation batch size."""
-
-    n_doe_iterations: int = field(validator=instance_of(int))
-    """The number of Design of Experiment iterations."""
-
-    n_mc_iterations: int = field(validator=instance_of(int))
-    """The number of Monte Carlo iterations."""
 
 
 @define(frozen=True)
@@ -83,17 +69,6 @@ class Benchmark(Generic[BenchmarkSettingsType], BenchmarkSerialization):
         )
 
         return Result(self.name, result, metadata)
-
-
-@define(frozen=True)
-class ConvergenceBenchmark(Benchmark):
-    """A class for defining convergence benchmarks."""
-
-    best_possible_result: float | None = field(default=None)
-    """The best possible result which can be achieved in the optimization process."""
-
-    optimal_function_inputs: list[dict[str, Any]] | None = field(default=None)
-    """An input that creates the best_possible_result."""
 
 
 # Register un-/structure hooks
